+++ packages/draw/src/constants/index.ts
@@ -3,3 +3,4 @@ export * from './pointer';
 export * from './image';
 export * from './theme';
 export * from './draw';
+export * from './swimlane';