+++ packages/draw/src/plugins/with-swimlane.ts
@@ -9,11 +9,10 @@ import {
 } from '@plait/core';
 import { PlaitDrawElement, PlaitSwimlane } from '../interfaces';
 import { buildSwimlaneTable } from '../utils/swimlane';
-import { PlaitTableBoard } from './with-table';
-import { PlaitTable } from '../interfaces/table';
 import { TableComponent } from '../table.component';
 import { withSwimlaneCreateByDrag, withSwimlaneCreateByDrawing } from './with-swimlane-create';
 import { getHitCell, setSelectedCells, isSingleSelectSwimlane, getSelectedSwimlane } from '../utils';
+import { PlaitBaseTable, PlaitTableBoard } from '../interfaces/table';
 
 export const withSwimlane = (board: PlaitTableBoard) => {
     const { drawElement, buildTable, pointerUp } = board;
@@ -25,7 +24,7 @@ export const withSwimlane = (board: PlaitTableBoard) => {
         return drawElement(context);
     };
 
-    board.buildTable = (element: PlaitTable) => {
+    board.buildTable = (element: PlaitBaseTable) => {
         if (PlaitDrawElement.isSwimlane(element)) {
             return buildSwimlaneTable(element as PlaitSwimlane);
         }