+++ packages/draw/src/utils/index.ts
@@ -10,3 +10,4 @@ export * from './common';
 export * from './snap-resizing';
 export * from './table';
 export * from './table-selected';
+export * from './swimlane';