+++ packages/draw/src/utils/table-selected.ts
@@ -1,9 +1,15 @@
 import { PlaitBoard, getSelectedElements, PlaitElement } from '@plait/core';
-import { PlaitTableElement, PlaitTable, PlaitTableCell } from '../interfaces/table';
+import { PlaitTableElement, PlaitTableCell, PlaitBaseTable, PlaitTable } from '../interfaces/table';
+import { PlaitDrawElement } from '../interfaces';
 
 export const isSingleSelectTable = (board: PlaitBoard) => {
     const selectedElements = getSelectedElements(board);
     return selectedElements && selectedElements.length === 1 && PlaitTableElement.isTable(selectedElements[0]);
+};
+
+export const isSingleSelectElementByTable = (board: PlaitBoard) => {
+    const selectedElements = getSelectedElements(board);
+    return selectedElements && selectedElements.length === 1 && PlaitDrawElement.isElementByTable(selectedElements[0]);
 };
 
 export const getSelectedTableElements = (board: PlaitBoard, elements?: PlaitElement[]) => {
@@ -11,16 +17,16 @@ export const getSelectedTableElements = (board: PlaitBoard, elements?: PlaitElem
     return selectedElements.filter(value => PlaitTableElement.isTable(value)) as PlaitTable[];
 };
 
-export const SELECTED_CELLS = new WeakMap<PlaitTable, PlaitTableCell[]>();
+export const SELECTED_CELLS = new WeakMap<PlaitBaseTable, PlaitTableCell[]>();
 
-export function getSelectedCells(element: PlaitTable) {
+export function getSelectedCells(element: PlaitBaseTable) {
     return SELECTED_CELLS.get(element);
 }
 
-export function setSelectedCells(element: PlaitTable, cells: PlaitTableCell[]) {
+export function setSelectedCells(element: PlaitBaseTable, cells: PlaitTableCell[]) {
     return SELECTED_CELLS.set(element, cells);
 }
 
-export function clearSelectedCells(element: PlaitTable) {
+export function clearSelectedCells(element: PlaitBaseTable) {
     return SELECTED_CELLS.delete(element);
 }